Rename to machinist.
This commit changes the project name, organization, and updates the API a
little bit:

 * The spire.ops is renamed to machinist.
 * SpireOps is renamed to DefaultOps
 * SpireOperatorNames is renamed to DefaultOperatorNames

That's it! I'm expecting a PR from Lars to set up some sbt-release magic, so
for now the "new" README.md is located at FUTURE.md to avoid confusing people.
As soon as I get a version released under this name/organization I will swap
that in.